@@ -12,46 +12,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~5.2.4",
-    "@angular/cdk": "~5.2.4",
-    "@angular/common": "~5.2.4",
-    "@angular/compiler": "~5.2.4",
-    "@angular/core": "~5.2.4",
-    "@angular/forms": "~5.2.4",
-    "@angular/http": "~5.2.4",
-    "@angular/material": "~5.2.0",
-    "@angular/material-moment-adapter": "~5.2.0",
-    "@angular/platform-browser": "~5.2.4",
-    "@angular/platform-browser-dynamic": "~5.2.4",
-    "@angular/router": "~5.2.4",
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.0.0",
+    "@angular/material-moment-adapter": "^6.2.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
     "bootstrap": "4.0.0-alpha.6",
     "core-js": "~2.4.1",
     "font-awesome": "~4.7.0",
     "jquery": "~3.3.1",
     "moment": "~2.20.1",
-    "rxjs": "~5.5.2",
-    "zone.js": "~0.8.20"
+    "rxjs": "^6.2.0",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.3",
-    "@angular/compiler-cli": "~5.2.4",
-    "@angular/language-service": "~5.2.4",
+    "@angular-devkit/build-angular": "~0.6.5",
+    "@angular/cli": "6.0.5",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/lodash": "^4.14.107",
-    "@types/node": "~6.0.100",
+    "@types/lodash": "^4.14.109",
+    "@types/node": "^6.0.111",
     "codelyzer": "~4.1.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "~1.4.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "~0.2.2",
-    "lodash": "^4.17.5",
+    "lodash": "^4.17.10",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.5.0"
+    "typescript": "~2.7.2"
   }
 }
